+++ packages/vaster_llama_ffi/lib/vaster_llama_ffi.dart
@@ -12,4 +12,7 @@ export 'src/bindings/llama_bindings.dart'
     show LlamaBindings, LlamaModelParams, LlamaContextParams, LlamaBatch;
 export 'src/llama_engine.dart'
     show LlamaEngine, LlamaStateIncompatibleException;
+export 'src/llama_ffi_kv_cache_controller.dart'
+    show LlamaFfiKvCacheController;
+export 'src/llama_ffi_vaster_model.dart' show LlamaFfiVasterModel;
 export 'src/llama_worker.dart' show LlamaWorker;